# api-gateway/api/docs/swagger.yaml
definitions:
...
  order.CartGetAllRes:
    properties:
      carts:
        items:
          $ref: '#/definitions/order.CartRes'
        type: array
    type: object
  order.CartGetByIdRes:
    properties:
      cart:
        $ref: '#/definitions/order.CartRes'
    type: object
  order.CartItemCreateReq:
    properties:
      cart_id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  order.CartItemGetAllRes:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/order.CartItemRes'
        type: array
    type: object
  order.CartItemGetByIdRes:
    properties:
      cart_item:
        $ref: '#/definitions/order.CartItemRes'
    type: object
  order.CartItemRes:
    properties:
      cart_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
  order.CartRes:
    properties:
      id:
        type: string
      total_amount:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /cart:
    post:
      consumes:
      - application/json
      description: Create a new cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create cart
      tags:
      - cart
  /cart/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a cart by ID
      produces:
      - application/json
      responses:
        "200":
          description: Cart deleted
          schema:
            type: string
        "400":
          description: Invalid event ID
          schema:
            type: string
        "404":
          description: Cart not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete cart
      tags:
      - cart
    get:
      consumes:
      - application/json
      description: Get a cart
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.CartGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Update an existing cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Cart not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update cart
      tags:
      - cart
  /cart/all:
    get:
      consumes:
      - application/json
      description: Get all carts
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.CartGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all carts
      tags:
      - cart
  /cart_item:
    post:
      consumes:
      - application/json
      description: Create a new cart
      parameters:
      - description: Cart data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/order.CartItemCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Cart item created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create cart
      tags:
      - cart item
  /cart_item/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a cart by ID
      parameters:
      - description: Cart ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart deleted
          schema:
            type: string
        "400":
          description: Invalid event ID
          schema:
            type: string
        "404":
          description: Cart not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete cart
      tags:
      - cart item
    get:
      consumes:
      - application/json
      description: Get a cart
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.CartItemGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - cart item
    put:
      consumes:
      - application/json
      description: Update an existing cart
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      - description: Quantity
        in: query
        name: quantity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cart updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Cart not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update cart
      tags:
      - cart item
  /cart_item/all:
    get:
      consumes:
      - application/json
      description: Get all carts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.CartItemGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all carts
      tags:
      - cart item
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"